lambdas/function/requirements.txt
google-cloud-storage
unittest
jsonpickle
aws-xray-sdk
boto3